Removido collections de transações antigas
-- Removido a collection antiga de transactions, além da sendPending e receivePending
-- Substituído a collection de transactions pela newTransactions, que foi renomeada para 'transaction'
